requirements.txt
gradio
gitpython
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
tqdm
